# src/main/java/se/sundsvall/casemanagement/api/model/OtherCaseDTO.java
package se.sundsvall.casemanagement.api.model;

import io.swagger.v3.oas.annotations.media.Schema;
import java.io.Serializable;
import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.ToString;
import lombok.experimental.SuperBuilder;

@EqualsAndHashCode(callSuper = true)
@ToString(callSuper = true)
@Data
@SuperBuilder(setterPrefix = "with")
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Other case model")
public class OtherCaseDTO extends CaseDTO implements Serializable {

	private static final long serialVersionUID = -1627503596853775503L;

	@Schema(description = "The facilities in the case")
	private List<FacilityDTO> facilities;
}