# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: ms-casemanagementservice
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "8.1"
servers:
  - url: http://localhost:63306
    description: Generated server url
...
paths:
  /{municipalityId}/cases/{externalCaseId}:
    put:
      tags:
        - Cases
      description: Update a case.
      operationId: putCase
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: externalCaseId
          in: path
          description: External case id
          required: true
          schema:
            type: string
          example: 1234
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDTO"
        required: true
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/cases:
    post:
      tags:
        - Cases
      description: Creates a case in ByggR or Ecos2 based on caseType. Also persists
        a connection between externalCaseId and the created case.
      operationId: postCases
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseResourceResponseDTO"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/CaseResourceResponseDTO"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{partyId}/statuses:
    get:
      tags:
        - Status
      description: Returns the case status for all cases where the specified party
        is involved.
      operationId: getStatusesByPartyId
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: partyId
          in: path
          description: Party id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStatusDTO"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStatusDTO"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/organization/{organizationNumber}/cases/status:
    get:
      tags:
        - Status
      description: Returns the latest status for each of the cases where the specified
        organization has the role "applicant".
      operationId: getStatusByOrgNr
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: organizationNumber
          in: path
          description: OrganizationNumber
          required: true
          schema:
            pattern: "^((18|19|20|21)\\d{6}|\\d{6})-(\\d{4})$"
            type: string
            description: Organization number with 10 or 12 digits.
            example: 20220622-2396
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStatusDTO"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStatusDTO"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/cases/{externalCaseId}/status:
    get:
      tags:
        - Status
      description: Returns the latest status for the case in the underlying system
        connected to the specified externalCaseId.
      operationId: getStatusByExternalCaseId
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: externalCaseId
          in: path
          description: External case id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStatusDTO"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/CaseStatusDTO"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
...
    AttachmentDTO:
      required:
        - extension
        - file
        - mimeType
        - name
      type: object
      properties:
        category:
          type: string
          description: The attachment category
          example: DOCUMENT
        name:
          type: string
          description: The attachment name
          example: Attachment name
        note:
          type: string
          description: A note on an attachment
          example: A note on an attachment.
        extension:
          type: string
          description: The file extension
          example: .pdf
        mimeType:
          type: string
          description: the mime type of the attachment
          example: application/pdf
        file:
          type: string
          description: Base64-encoded file (plain text)
          format: byte
          example: ZEdWemRBPT0=
        extraParameters:
          type: object
          additionalProperties:
            type: string
            description: Extra parameters for the attachment
          description: Extra parameters for the attachment
      description: Attachment model
    ByggRCaseDTO:
      required:
        - attachments
        - caseType
        - externalCaseId
        - facilities
        - stakeholders
      type: object
      description: ByggR-cases
      allOf:
        - $ref: "#/components/schemas/CaseDTO"
        - type: object
          properties:
            diaryNumber:
              type: string
              description: The case diary number
              example: 2021-1234
            facilities:
              type: array
              description: The facilities in the case
              items:
                $ref: "#/components/schemas/FacilityDTO"
    CaseDTO:
      required:
        - attachments
        - caseType
        - externalCaseId
        - stakeholders
      type: object
      properties:
        externalCaseId:
          type: string
          description: Case ID from the client.
          example: caa230c6-abb4-4592-ad9a-34e263c2787b
        caseType:
          type: string
          description: The case type
          example: NYBYGGNAD_ANSOKAN_OM_BYGGLOV
        description:
          type: string
          description: Some description of the case.
          example: En fritextbeskrivning av case.
        caseTitleAddition:
          type: string
          description: "Additions to the case title. Right now only applicable to\
            \ cases of CaseType: NYBYGGNAD_ANSOKAN_OM_BYGGLOV."
          example: "Eldstad/rökkanal, Skylt"
        stakeholders:
          type: array
          description: The stakeholders in the case
          items:
            oneOf:
              - $ref: "#/components/schemas/OrganizationDTO"
              - $ref: "#/components/schemas/PersonDTO"
        attachments:
          type: array
          description: The attachments in the case
          items:
            $ref: "#/components/schemas/AttachmentDTO"
        extraParameters:
          type: object
          additionalProperties:
            type: string
            description: Extra parameters for the case.
          description: Extra parameters for the case.
      description: Base case model
      example:
        externalCaseId: e19981ad-34b2-4e14-88f5-133f61ca85aa
        caseType: NYBYGGNAD_ANSOKAN_OM_BYGGLOV
        facilities:
          - facilityType: STOREHOUSE
            address:
              addressCategories:
                - VISITING_ADDRESS
              propertyDesignation: SUNDSVALL BALDER 2
        stakeholders:
          - type: ORGANIZATION
            roles:
              - APPLICANT
              - PAYMENT_PERSON
            organizationName: Testorganisationen
            organizationNumber: 123456-1234
          - type: PERSON
            roles:
              - CONTACT_PERSON
            firstName: Test
            lastName: Testsson
            personId: e19981ad-34b2-4e14-88f5-133f61ca85aa
        attachments:
          - category: ANS
            name: Some_name_2022-03-07
            extension: .pdf
            file: dGVzdA==
      discriminator:
        propertyName: caseType
      oneOf:
        - $ref: "#/components/schemas/ByggRCaseDTO"
        - $ref: "#/components/schemas/EcosCaseDTO"
        - $ref: "#/components/schemas/OtherCaseDTO"
...
    EcosCaseDTO:
      required:
        - attachments
        - caseType
        - externalCaseId
        - facilities
        - stakeholders
      type: object
      description: Ecos2-case model class
      allOf:
        - $ref: "#/components/schemas/CaseDTO"
        - type: object
          properties:
            startDate:
              type: string
              description: Start date for the business.
              format: date
              example: 2022-01-01
            endDate:
              type: string
              description: End date of the business if it is time-limited.
              format: date
              example: 2022-06-01
            facilities:
              maxItems: 1
              minItems: 1
              type: array
              description: The facilities in the case
              items:
                $ref: "#/components/schemas/FacilityDTO"
...
    OrganizationDTO:
      required:
        - organizationName
        - organizationNumber
        - type
      type: object
      description: Organization model
      allOf:
        - $ref: "#/components/schemas/StakeholderDTO"
        - type: object
          properties:
            organizationName:
              type: string
              description: Organization name
              example: Sundsvalls testfabrik
            organizationNumber:
              type: string
              description: Organization number with 10 or 12 digits.
              example: 20220622-2396
            authorizedSignatory:
              type: string
              description: The authorized signatory
              example: Test Testorsson
    OtherCaseDTO:
      required:
        - attachments
        - caseType
        - externalCaseId
        - stakeholders
      type: object
      description: Other case model
      allOf:
        - $ref: "#/components/schemas/CaseDTO"
        - type: object
          properties:
            facilities:
              type: array
              description: The facilities in the case
              items:
                $ref: "#/components/schemas/FacilityDTO"
    PersonDTO:
      required:
        - firstName
        - lastName
        - personId
        - type
      type: object
      description: Person model
      allOf:
        - $ref: "#/components/schemas/StakeholderDTO"
        - type: object
          properties:
            firstName:
              type: string
              description: The person's first name
              example: Test
            lastName:
              type: string
              description: The person's last name
              example: Testorsson
            personId:
              type: string
              description: The person's id
              example: 3ed5bc30-6308-4fd5-a5a7-78d7f96f4438
            personalNumber:
              type: string
              description: The person's personal number
              readOnly: true
              example: 20000101-1234
    StakeholderDTO:
      required:
        - type
      type: object
      properties:
        type:
          type: string
          description: Stakeholder type
          example: PERSON
          enum:
            - PERSON
            - ORGANIZATION
        roles:
          type: array
          description: An stakeholder can have one or more roles. Please note that
            INVOICE_RECIPENT is deprecated and should not be used. Use INVOICE_RECIPIENT
            instead.
          items:
            type: string
            description: An stakeholder can have one or more roles. Please note that
              INVOICE_RECIPENT is deprecated and should not be used. Use INVOICE_RECIPIENT
              instead.
        phoneNumber:
          type: string
          description: Stakeholder phone number
          example: "060123456"
        emailAddress:
          type: string
          description: Stakeholder emailaddress
          example: test.testorsson@sundsvall.se
        cellphoneNumber:
          type: string
          description: Stakeholder cellphone number
          example: "0701234567"
        addresses:
          type: array
          description: An stakeholder may have one or more addresses. For example
            one POSTAL_ADDRESS and another INVOICE_ADDRESS.
          items:
            $ref: "#/components/schemas/AddressDTO"
        extraParameters:
          type: object
          additionalProperties:
            type: string
            description: The stakeholder's billing address
          description: The stakeholder's billing address
      description: Stakeholder model
      discriminator:
        propertyName: type
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        title:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/StatusType"
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        title:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/StatusType"
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CaseResourceResponseDTO:
      type: object
      properties:
        caseId:
          type: string
          description: The case id
          example: "1234"
      description: Case resource response
    CaseStatusDTO:
      type: object
      properties:
        system:
          type: string
          description: System type
          example: BYGGR
          enum:
            - BYGGR
            - ECOS
            - CASE_DATA
            - ALKT
        caseType:
          type: string
          description: The case type
          example: BYGGLOV
        externalCaseId:
          type: string
          description: The external case id
          example: caa230c6-abb4-4592-ad9a-34e263c2787d
        caseId:
          type: string
          description: The case id
          example: BYGG 2022-000100
        status:
          type: string
          description: Case status
          example: Pågående
        serviceName:
          type: string
          description: Service name
          example: Nybyggnad - Ansökan om bygglov
        timestamp:
          type: string
          description: The timestamp
          format: date-time
      description: Case status model
...
  securitySchemes: { }